refactor(test): refactor imports in load test scripts
