+++ __tests__/performance/load/payment.load.js
@@ -1,10 +1,11 @@
+// Jovin Ang Yusheng, A0273460H
 import http from "k6/http";
 import { sleep } from "k6";
 import { createLoadOptions } from "./configs/thresholds.js";
 import { buildAuthHeaders, getLoadUserPool, loginUser, pickUserForVu } from "./helpers/auth.js";
-import { getBaseUrl, getNumberEnv, getOptionalEnv } from "./helpers/env.js";
-import { recordTransaction, trackResponse } from "./helpers/metrics.js";
-import { buildCart } from "./helpers/payloads.js";
+import { getBaseUrl, getNumberEnv, getOptionalEnv } from "../common/k6/env.js";
+import { recordTransaction, trackResponse } from "../common/k6/metrics.js";
+import { buildCart } from "../common/k6/payloads.js";
 
 /*
  * Load Test — Checkout & Payment
