+++ __tests__/performance/load/scripts/runAllLoadFlows.js
@@ -1,3 +1,4 @@
+// Jovin Ang Yusheng, A0273460H
 import { spawnSync } from "child_process";
 import path from "path";
 import { getProjectRoot } from "./loadDataManager.js";
